src/requirements.txt
google-cloud-pubsub
google-cloud
requests
dataclasses-json
ndjson
functions-framework
python-dateutil